- fix main view controller not attached in ios causing shared ui not to progress  - delete unnecessary platform spread controllers + di: we can assume    default main controller is client controller
