+++ bisqapps/androidClient/src/androidMain/kotlin/network/bisq/mobile/client/MainApplication.kt
@@ -1,7 +1,6 @@
 package network.bisq.mobile.client
 
 import android.app.Application
-import network.bisq.mobile.client.di.androidClientModule
 import network.bisq.mobile.client.di.clientModule
 import network.bisq.mobile.domain.di.domainModule
 import network.bisq.mobile.presentation.di.presentationModule
@@ -15,7 +14,7 @@ class MainApplication: Application() {
 
         startKoin {
             androidContext(this@MainApplication)
-            modules(listOf(domainModule, presentationModule, clientModule, androidClientModule))
+            modules(listOf(domainModule, presentationModule, clientModule))
         }
     }
 }
