+++ bisqapps/shared/presentation/src/commonMain/kotlin/network/bisq/mobile/presentation/MainPresenter.kt
@@ -5,12 +5,13 @@ import co.touchlab.kermit.Logger
 import kotlinx.coroutines.flow.*
 import network.bisq.mobile.android.node.BuildNodeConfig
 import network.bisq.mobile.client.shared.BuildConfig
+import network.bisq.mobile.domain.data.repository.main.bootstrap.ApplicationBootstrapFacade
 import network.bisq.mobile.presentation.ui.AppPresenter
 
 /**
  * Main Presenter as an example of implementation for now.
  */
-open class MainPresenter() : BasePresenter(null), AppPresenter {
+open class MainPresenter(private val applicationBootstrapFacade: ApplicationBootstrapFacade) : BasePresenter(null), AppPresenter {
     lateinit var navController: NavHostController
         private set
 
@@ -22,6 +23,16 @@ open class MainPresenter() : BasePresenter(null), AppPresenter {
     // Observable state
     private val _isContentVisible = MutableStateFlow(false)
     override val isContentVisible: StateFlow<Boolean> = _isContentVisible
+
+   private var applicationServiceInited = false
+    override fun onViewAttached() {
+        super.onViewAttached()
+
+        if (!applicationServiceInited) {
+            applicationServiceInited = true
+            applicationBootstrapFacade.initialize()
+        }
+    }
 
     // passthrough example
 //    private val _greetingText: StateFlow<String> = stateFlowFromRepository(
