--- bisqapps/shared/presentation/src/iosMain/kotlin/network/bisq/mobile/presentation/IosClientMainPresenter.kt
@@ -1,19 +1,0 @@
-package network.bisq.mobile.presentation
-
-import network.bisq.mobile.domain.data.repository.GreetingRepository
-import network.bisq.mobile.domain.data.repository.main.bootstrap.ApplicationBootstrapFacade
-
-@Suppress("UNCHECKED_CAST")
-class IosClientMainPresenter(
-    private val applicationBootstrapFacade: ApplicationBootstrapFacade
-) : MainPresenter() {
-    var applicationServiceInited = false
-    override fun onViewAttached() {
-        super.onViewAttached()
-
-        if (!applicationServiceInited) {
-            applicationServiceInited = true
-            applicationBootstrapFacade.initialize()
-        }
-    }
-}
