+++ bisqapps/shared/presentation/src/iosMain/kotlin/network/bisq/mobile/presentation/di/DependenciesProviderHelper.kt
@@ -14,7 +14,7 @@ class DependenciesProviderHelper {
 
     fun initKoin() {
         val instance = startKoin {
-            modules(listOf(domainModule, presentationModule, clientModule, iosClientModule))
+            modules(listOf(domainModule, presentationModule, clientModule))
         }
 
         koin = instance.koin
